requirements.txt
mpneuralnetwork
numpy
scipy
pillow
ruff
